Decompose 5 large frontend components into focused sub-modules
Break up the largest frontend components to improve maintainability:

- StyledContainers.tsx (2,115→12 lines): Extract 9 styled-component
  files into styled/ directory with barrel re-export
- SystemSettings.tsx (2,616→1,108 lines): Extract GraphQL operations,
  types, styles, and 4 sub-components into system_settings/
- CorpusChat.tsx (2,347→1,346 lines): Extract styles, ApprovalModal,
  and ConversationListView into corpus_chat/
- ChatTray.tsx (2,215→1,772 lines): Extract ApprovalOverlay,
  ConversationListView, and chatUtils into sibling files
- DocumentKnowledgeBase.tsx (3,363→2,322 lines): Extract styles,
  useZoomManager hook, RightPanelContent, DocumentModals, and
  ContextBar into document_kb/
- Consolidate duplicate WebSocket types (WebSocketSources, MessageData,
  ContextStatus, CompactionNotice) into components/chat/types.ts

All existing imports remain backward-compatible via barrel re-exports.

Closes #977
