+++ frontend/src/components/corpuses/CorpusChat.tsx
@@ -24,22 +24,8 @@ import React, {
 } from "react";
 import { useLazyQuery, useQuery, useReactiveVar } from "@apollo/client";
 import { AnimatePresence, motion } from "framer-motion";
-import { formatDistanceToNow } from "date-fns";
-import {
-  AlertCircle,
-  ArrowLeft,
-  Calendar,
-  Plus,
-  Search,
-  Send,
-  X,
-  Home,
-  CheckCircle,
-} from "lucide-react";
+import { AlertCircle, ArrowLeft, Send, Home } from "lucide-react";
 import { Button, Loader } from "semantic-ui-react";
-import styled from "styled-components";
-import { color } from "../../theme/colors";
-import { MessageCircle } from "lucide-react";
 import { CONVERSATION_TYPE } from "../../assets/configurations/constants";
 
 import {
@@ -53,98 +39,48 @@ import {
 
 import {
   ErrorContainer,
-  ConversationGrid,
-  ChatInputContainer,
-  ChatInput,
-  SendButton,
   ErrorMessage,
   ConnectionStatus,
-  FilterContainer,
 } from "../knowledge_base/document/ChatContainers";
 
-import { authToken, userObj, showQueryViewState } from "../../graphql/cache";
-import {
-  DatePickerExpanded,
-  ExpandingInput,
-  IconButton,
-} from "../knowledge_base/document/FilterContainers";
+import { authToken, userObj } from "../../graphql/cache";
+import { IconButton } from "../knowledge_base/document/FilterContainers";
 import {
   useChatSourceState,
   mapWebSocketSourcesToChatMessageSources,
   ChatMessageSource,
 } from "../annotator/context/ChatSourceAtom";
-import { MultipageAnnotationJson } from "../types";
-import { FetchMoreOnVisible } from "../widgets/infinite_scroll/FetchMoreOnVisible";
 import {
   ChatMessage,
   ChatMessageProps,
   TimelineEntry,
 } from "../widgets/chat/ChatMessage";
 import { getCorpusQueryWebSocket } from "../chat/get_websockets";
-import { MOBILE_VIEW_BREAKPOINT } from "../../assets/configurations/constants";
-import useWindowDimensions from "../hooks/WindowDimensionHook";
+import type {
+  WebSocketSources,
+  MessageData,
+  ContextStatus,
+  CompactionNotice,
+} from "../chat/types";
 
-/**
- * A helper interface representing the properties of data included in websocket messages,
- * specifically any source annotations or label metadata.
- */
-interface WebSocketSources {
-  page: number;
-  json: { start: number; end: number } | MultipageAnnotationJson;
-  annotation_id: number;
-  label: string;
-  label_id: number;
-  rawText: string;
-}
-
-/**
- * An interface for the full websocket message structure.
- * Includes the type of content (async start, streaming content, final, etc.),
- * the actual text content, and optional annotation data.
- */
-interface MessageData {
-  type:
-    | "ASYNC_START"
-    | "ASYNC_CONTENT"
-    | "ASYNC_FINISH"
-    | "SYNC_CONTENT"
-    | "ASYNC_THOUGHT"
-    | "ASYNC_SOURCES"
-    | "ASYNC_APPROVAL_NEEDED"
-    | "ASYNC_APPROVAL_RESULT"
-    | "ASYNC_RESUME"
-    | "ASYNC_ERROR";
-  content: string;
-  data?: {
-    sources?: WebSocketSources[];
-    timeline?: TimelineEntry[];
-    message_id?: string;
-    tool_name?: string;
-    args?: any;
-    pending_tool_call?: {
-      name: string;
-      arguments: any;
-      tool_call_id?: string;
-    };
-    [key: string]: any;
-  };
-}
-
-/**
- * Context status metadata from the backend (token usage, compaction info).
- */
-interface ContextStatus {
-  used_tokens: number;
-  context_window: number;
-  was_compacted: boolean;
-  tokens_before_compaction: number;
-}
-
-interface CompactionNotice {
-  tokensBefore: number;
-  tokensAfter: number;
-  contextWindow: number;
-}
+import {
+  ChatContainer,
+  ConversationIndicator,
+  ChatNavigationHeader,
+  BackButton,
+  NavigationTitle,
+  MessagesArea,
+  MessageWrapper,
+  LatestMessageIndicator,
+  ProcessingIndicator,
+  ChatInputWrapper,
+  EnhancedChatInputContainer,
+  EnhancedChatInput,
+  EnhancedSendButton,
+  InputRow,
+} from "./corpus_chat/styles";
+import { ApprovalModal, PendingApproval } from "./corpus_chat/ApprovalModal";
+import { ConversationListView } from "./corpus_chat/ConversationListView";
 
 /**
  * CorpusChat props definition.
@@ -170,622 +106,6 @@ interface CorpusChatProps {
   onSourceNavigate?: (source: ChatMessageSource) => void;
 }
 
-// Add these styled components near your other styled components
-const ConversationHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1.25rem 1.5rem;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.08);
-  background: rgba(255, 255, 255, 0.98);
-  backdrop-filter: blur(8px);
-  position: sticky;
-  top: 0;
-  z-index: 10;
-
-  h2 {
-    font-size: 1.5rem;
-    font-weight: 600;
-    color: #1a202c;
-    margin: 0;
-  }
-
-  .actions {
-    display: flex;
-    gap: 0.75rem;
-  }
-
-  @media (max-width: 768px) {
-    padding: 1rem;
-
-    h2 {
-      font-size: 1.25rem;
-    }
-
-    .actions button {
-      padding: 0.5rem 0.75rem !important;
-      font-size: 0.85rem !important;
-    }
-  }
-`;
-
-const EnhancedConversationGrid = styled(ConversationGrid)`
-  display: flex;
-  flex-direction: column;
-  gap: 0;
-  padding: 0;
-  width: 100%;
-  overflow-y: auto;
-  background: ${color.N2};
-`;
-
-const EnhancedConversationCard = styled(motion.div)`
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  padding: 1rem 1.5rem;
-  background: ${color.N1};
-  border-bottom: 1px solid ${color.N4};
-  cursor: pointer;
-  transition: all 0.15s ease;
-  position: relative;
-
-  &:hover {
-    background: ${color.G1};
-  }
-
-  &:active {
-    background: ${color.G2};
-  }
-`;
-
-const ChatItemIcon = styled.div`
-  width: 40px;
-  height: 40px;
-  border-radius: 10px;
-  background: linear-gradient(135deg, ${color.B2} 0%, ${color.B3} 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-
-  svg {
-    width: 20px;
-    height: 20px;
-    color: ${color.B7};
-  }
-`;
-
-const ChatItemContent = styled.div`
-  flex: 1;
-  min-width: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-`;
-
-const ChatItemTitle = styled.div`
-  font-size: 0.9375rem;
-  font-weight: 600;
-  color: ${color.N10};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-const ChatItemMeta = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  font-size: 0.8125rem;
-  color: ${color.N6};
-`;
-
-const MessageCountBadge = styled.div`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 24px;
-  height: 24px;
-  padding: 0 8px;
-  border-radius: 12px;
-  background: ${color.B4};
-  color: white;
-  font-size: 0.75rem;
-  font-weight: 600;
-  flex-shrink: 0;
-`;
-
-const EnhancedFilterContainer = styled(FilterContainer)`
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding: 0.75rem 1.5rem;
-  background: ${color.N1};
-  border-bottom: 1px solid ${color.N4};
-  margin: 0;
-  border-radius: 0;
-  position: sticky;
-  top: 0;
-  z-index: 10;
-`;
-
-const NewChatButton = styled(motion.button)`
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 6px;
-  background: ${color.G6};
-  color: white;
-  font-size: 0.875rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s;
-  white-space: nowrap;
-  flex-shrink: 0;
-
-  &:hover {
-    background: ${color.G7};
-  }
-
-  svg {
-    width: 16px;
-    height: 16px;
-  }
-
-  @media (max-width: 480px) {
-    padding: 0.5rem;
-    span {
-      display: none;
-    }
-  }
-`;
-
-const ToolbarDivider = styled.div`
-  width: 1px;
-  height: 24px;
-  background: ${color.N4};
-  margin: 0 0.25rem;
-  flex-shrink: 0;
-`;
-
-const EmptyStateContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 4rem 2rem;
-  text-align: center;
-  flex: 1;
-  background: ${color.N2};
-`;
-
-const EmptyStateIcon = styled.div`
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  background: linear-gradient(135deg, ${color.G1} 0%, ${color.G2} 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 1.5rem;
-
-  svg {
-    width: 36px;
-    height: 36px;
-    color: ${color.G6};
-  }
-`;
-
-const EmptyStateTitle = styled.h3`
-  font-size: 1.25rem;
-  font-weight: 700;
-  margin: 0 0 0.5rem;
-  color: ${color.N10};
-  letter-spacing: -0.025em;
-`;
-
-const EmptyStateDescription = styled.p`
-  color: ${color.N7};
-  max-width: 360px;
-  margin: 0 0 1.5rem;
-  font-size: 0.9375rem;
-  line-height: 1.5;
-`;
-
-const EmptyStateButton = styled(motion.button)`
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 8px;
-  background: ${color.G6};
-  color: white;
-  font-size: 0.9375rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    background: ${color.G7};
-    transform: translateY(-1px);
-  }
-
-  svg {
-    width: 18px;
-    height: 18px;
-  }
-`;
-
-// Update the ChatContainer styled component
-const ChatContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-  overflow: hidden;
-  background: #f8fafc;
-  position: relative;
-  margin: 0;
-  padding: 0;
-  border-radius: 0;
-  flex: 1;
-  min-height: 0; /* Important for flex children */
-  max-height: 100%; /* Never exceed parent's height */
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    /* Removed position: fixed to prevent covering parent navigation and bottom nav */
-    /* Container now works as a normal flex child */
-    height: 100%; /* Fill parent container */
-    max-height: 100%; /* Don't exceed parent's height */
-  }
-`;
-
-const ConversationIndicator = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-  overflow: hidden;
-  position: relative;
-  min-height: 0; /* Important for flex children to properly overflow */
-  flex: 1;
-  max-height: 100%; /* Never exceed parent's height */
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    height: 100%; /* Fill parent container */
-    max-height: 100%; /* Don't exceed parent's height */
-  }
-`;
-
-// Add a styled component for the input wrapper
-const ChatInputWrapper = styled.div`
-  flex-shrink: 0;
-  width: 100%;
-  background: rgba(255, 255, 255, 0.98);
-  backdrop-filter: blur(10px);
-  border-top: 1px solid rgba(0, 0, 0, 0.08);
-  box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.04);
-  position: sticky;
-  bottom: 0;
-  z-index: 50; /* Ensure it's above other content */
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    position: sticky;
-    bottom: 0; /* Back to bottom - FAB is now compact and out of the way */
-    /* Ensure it stays at bottom even when keyboard appears */
-    transform: translateZ(0); /* Force GPU acceleration */
-  }
-`;
-
-// Add new styled component for processing indicator
-const ProcessingIndicator = styled(motion.div)`
-  display: inline-flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding: 0.875rem 1.5rem;
-  background: linear-gradient(135deg, #f0f7ff 0%, #e6f2ff 100%);
-  color: #4a90e2;
-  border-radius: 24px;
-  font-weight: 500;
-  font-size: 0.9375rem;
-  position: relative;
-  overflow: hidden;
-  border: 1px solid #d4e3f4;
-  box-shadow: 0 2px 8px rgba(74, 144, 226, 0.15);
-
-  &::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: -100%;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(
-      90deg,
-      transparent 0%,
-      rgba(74, 144, 226, 0.15) 50%,
-      transparent 100%
-    );
-    animation: shimmer 2s infinite;
-  }
-
-  @keyframes shimmer {
-    0% {
-      left: -100%;
-    }
-    100% {
-      left: 100%;
-    }
-  }
-
-  .pulse-dot {
-    width: 6px;
-    height: 6px;
-    background: #4a90e2;
-    border-radius: 50%;
-    animation: pulse 1.5s ease-in-out infinite;
-    box-shadow: 0 0 4px rgba(74, 144, 226, 0.4);
-  }
-
-  @keyframes pulse {
-    0%,
-    100% {
-      transform: scale(1);
-      opacity: 0.8;
-    }
-    50% {
-      transform: scale(1.5);
-      opacity: 0.4;
-    }
-  }
-
-  /* Loader color override */
-  .ui.loader {
-    &:after {
-      border-color: #4a90e2 transparent transparent !important;
-    }
-  }
-`;
-
-// Enhance the ChatInputContainer for better mobile experience
-const EnhancedChatInputContainer = styled(ChatInputContainer)<{
-  $disabled?: boolean;
-}>`
-  padding: 1.25rem 1.5rem;
-  background: rgba(255, 255, 255, 0.98);
-  backdrop-filter: blur(20px);
-  border-top: 1px solid rgba(0, 0, 0, 0.05);
-  box-shadow: 0 -4px 16px rgba(0, 0, 0, 0.04);
-  flex-direction: column;
-  align-items: stretch;
-
-  /* When disabled (i.e. assistant is processing) */
-  ${(props) =>
-    props.$disabled &&
-    `
-      opacity: 0.6;        /* Visually indicate inactive state */
-    `}
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    padding: 1rem;
-    gap: 0.75rem;
-  }
-`;
-
-// Enhance the chat messages area for mobile
-const MessagesArea = styled.div<{ $isProcessing?: boolean }>`
-  flex: 1;
-  overflow-y: auto;
-  overflow-x: hidden;
-  padding: 1.5rem;
-  background: linear-gradient(to bottom, #f8fafc 0%, #ffffff 100%);
-  min-height: 0;
-  max-height: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  transition: background 0.3s ease;
-
-  ${(props) =>
-    props.$isProcessing &&
-    `
-    background: linear-gradient(to bottom, #f0f7ff 0%, #f8fbff 100%);
-    animation: subtleGlow 2s ease-in-out infinite;
-    
-    @keyframes subtleGlow {
-      0%, 100% {
-        background: linear-gradient(to bottom, #f0f7ff 0%, #f8fbff 100%);
-      }
-      50% {
-        background: linear-gradient(to bottom, #e6f2ff 0%, #f0f7ff 100%);
-      }
-    }
-  `}
-
-  /* Custom scrollbar */
-  &::-webkit-scrollbar {
-    width: 8px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background: #f1f5f9;
-    border-radius: 4px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background: #cbd5e1;
-    border-radius: 4px;
-
-    &:hover {
-      background: #94a3b8;
-    }
-  }
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    padding: 1rem;
-  }
-`;
-
-// Input row wrapper to keep input and send button together
-const InputRow = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.875rem;
-  width: 100%;
-`;
-
-// Enhance the chat input for better mobile experience
-const EnhancedChatInput = styled(ChatInput)`
-  background: #f8fafc;
-  border: 2px solid #e2e8f0;
-  border-radius: 12px;
-  padding: 0.875rem 1.25rem;
-  font-size: 0.9375rem;
-  transition: all 0.2s ease;
-  flex: 1;
-  min-height: 48px;
-
-  &:focus {
-    background: white;
-    border-color: #4299e1;
-    box-shadow: 0 0 0 3px rgba(66, 153, 225, 0.1);
-  }
-
-  &:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-  }
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    font-size: 0.875rem;
-    padding: 0.75rem 1rem;
-    min-height: 44px;
-  }
-`;
-
-// Enhanced send button that matches input height
-const EnhancedSendButton = styled(SendButton)`
-  width: 48px;
-  height: 48px;
-  align-self: center; /* Override the flex-end from base component */
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    width: 44px;
-    height: 44px;
-  }
-`;
-
-// Add a new styled component for latest message indicator
-const LatestMessageIndicator = styled(motion.div)`
-  position: absolute;
-  left: 0;
-  width: 4px;
-  height: 100%;
-  background: linear-gradient(to bottom, #4299e1, #3182ce);
-  border-radius: 0 4px 4px 0;
-`;
-
-// Add a pulsing dot for new messages
-const NewMessageDot = styled(motion.div)`
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-  width: 12px;
-  height: 12px;
-  background: #ef4444;
-  border-radius: 50%;
-  box-shadow: 0 0 0 4px rgba(239, 68, 68, 0.2);
-
-  &::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: #ef4444;
-    border-radius: 50%;
-    animation: pulse 2s infinite;
-  }
-
-  @keyframes pulse {
-    0% {
-      transform: scale(1);
-      opacity: 1;
-    }
-    50% {
-      transform: scale(1.5);
-      opacity: 0.3;
-    }
-    100% {
-      transform: scale(1);
-      opacity: 1;
-    }
-  }
-`;
-
-// Add message wrapper for better styling
-const MessageWrapper = styled(motion.div)<{ isLatest?: boolean }>`
-  position: relative;
-  margin-bottom: 1.5rem;
-
-  ${(props) =>
-    props.isLatest &&
-    `
-    & > * {
-      box-shadow: 0 4px 12px rgba(66, 153, 225, 0.15);
-      border: 1px solid rgba(66, 153, 225, 0.2);
-    }
-  `}
-`;
-
-// Mobile navigation header styled components
-const ChatNavigationHeader = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  padding: 1rem;
-  background: white;
-  border-bottom: 1px solid #e2e8f0;
-  position: sticky;
-  top: 0;
-  z-index: 100; /* High z-index to ensure it's always visible */
-  flex-shrink: 0; /* Prevent compression */
-
-  @media (max-width: ${MOBILE_VIEW_BREAKPOINT}px) {
-    position: -webkit-sticky; /* For Safari */
-    position: sticky;
-    padding: 0.75rem 1rem;
-  }
-`;
-
-const BackButton = styled(motion.button)`
-  background: #f8fafc;
-  border: 1px solid #e2e8f0;
-  border-radius: 8px;
-  padding: 0.5rem;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.2s ease;
-
-  &:hover {
-    background: #e2e8f0;
-  }
-`;
-
-const NavigationTitle = styled.span`
-  flex: 1;
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: #1a202c;
-`;
-
 /**
  * CorpusChat component provides:
  * 1) Initial user selection of either creating a new conversation or loading an existing one,
@@ -807,10 +127,6 @@ export const CorpusChat: React.FC<CorpusChatProps> = ({
   onViewModeChange,
   onSourceNavigate,
 }) => {
-  // Window dimensions for responsive layout
-  const { width } = useWindowDimensions();
-  const use_mobile_layout = width <= MOBILE_VIEW_BREAKPOINT;
-
   // Chat state
   const [isNewChat, setIsNewChat] = useState(forceNewChat);
   const [newMessage, setNewMessage] = useState("");
@@ -849,12 +165,6 @@ export const CorpusChat: React.FC<CorpusChatProps> = ({
   const [createdAtGte, setCreatedAtGte] = useState<string>("");
   const [createdAtLte, setCreatedAtLte] = useState<string>("");
 
-  // For dynamic display of filters
-  const [showSearch, setShowSearch] = useState(false);
-  const [showDatePicker, setShowDatePicker] = useState(false);
-  const searchInputRef = useRef<HTMLInputElement>(null);
-  const datePickerRef = useRef<HTMLDivElement>(null);
-
   // Context status (token usage, compaction info)
   const [contextStatus, setContextStatus] = useState<ContextStatus | null>(
     null
@@ -863,14 +173,8 @@ export const CorpusChat: React.FC<CorpusChatProps> = ({
     useState<CompactionNotice | null>(null);
 
   // Approval gate state (mirrors ChatTray)
-  const [pendingApproval, setPendingApproval] = useState<{
-    messageId: string;
-    toolCall: {
-      name: string;
-      arguments: any;
-      tool_call_id?: string;
-    };
-  } | null>(null);
+  const [pendingApproval, setPendingApproval] =
+    useState<PendingApproval | null>(null);
   const [showApprovalModal, setShowApprovalModal] = useState<boolean>(false);
 
   /**
@@ -974,27 +278,6 @@ export const CorpusChat: React.FC<CorpusChatProps> = ({
     }, 500);
     return () => clearTimeout(timer);
   }, [titleFilter]);
-
-  // Hide search or date picker if user clicks outside
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        searchInputRef.current &&
-        !searchInputRef.current.contains(event.target as Node)
-      ) {
-        setShowSearch(false);
-      }
-      if (
-        datePickerRef.current &&
-        !datePickerRef.current.contains(event.target as Node)
-      ) {
-        setShowDatePicker(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, []);
 
   // Combine serverMessages + local chat for final display
   const combinedMessages = [...serverMessages, ...chat];
@@ -2035,325 +1318,29 @@ export const CorpusChat: React.FC<CorpusChatProps> = ({
             </motion.div>
           ) : (
             // CONVERSATION MENU VIEW
-            <motion.div
-              key="conversation-menu"
-              style={{
-                width: "100%",
-                height: "100%",
-                maxHeight: "100%",
-                display: "flex",
-                flexDirection: "column",
-                overflowY: "auto",
-                flex: 1,
-              }}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-              transition={{ duration: 0.3 }}
-            >
-              <EnhancedFilterContainer>
-                <NewChatButton
-                  onClick={startNewChat}
-                  whileHover={{ scale: 1.02 }}
-                  whileTap={{ scale: 0.98 }}
-                >
-                  <Plus size={16} />
-                  <span>New Chat</span>
-                </NewChatButton>
-
-                <ToolbarDivider />
-
-                <AnimatePresence>
-                  {showSearch && (
-                    <ExpandingInput
-                      initial={{ width: 0, opacity: 0 }}
-                      animate={{ width: "auto", opacity: 1 }}
-                      exit={{ width: 0, opacity: 0 }}
-                      ref={searchInputRef}
-                    >
-                      <input
-                        className="expanded"
-                        placeholder="Search chats..."
-                        value={titleFilter}
-                        onChange={(e) => setTitleFilter(e.target.value)}
-                        autoFocus
-                      />
-                    </ExpandingInput>
-                  )}
-                </AnimatePresence>
-
-                <IconButton
-                  onClick={() => setShowSearch(!showSearch)}
-                  $isActive={!!titleFilter}
-                  whileTap={{ scale: 0.95 }}
-                  title="Search"
-                >
-                  <Search />
-                </IconButton>
-
-                <IconButton
-                  onClick={() => setShowDatePicker(!showDatePicker)}
-                  $isActive={!!(createdAtGte || createdAtLte)}
-                  whileTap={{ scale: 0.95 }}
-                  title="Filter by date"
-                >
-                  <Calendar />
-                </IconButton>
-
-                <AnimatePresence>
-                  {showDatePicker && (
-                    <DatePickerExpanded
-                      ref={datePickerRef}
-                      initial={{ opacity: 0, y: -10 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      exit={{ opacity: 0, y: -10 }}
-                    >
-                      <input
-                        type="date"
-                        value={createdAtGte}
-                        onChange={(e) => setCreatedAtGte(e.target.value)}
-                        placeholder="Start Date"
-                      />
-                      <input
-                        type="date"
-                        value={createdAtLte}
-                        onChange={(e) => setCreatedAtLte(e.target.value)}
-                        placeholder="End Date"
-                      />
-                    </DatePickerExpanded>
-                  )}
-                </AnimatePresence>
-
-                {(titleFilter || createdAtGte || createdAtLte) && (
-                  <IconButton
-                    onClick={() => {
-                      setTitleFilter("");
-                      setCreatedAtGte("");
-                      setCreatedAtLte("");
-                      setShowSearch(false);
-                      setShowDatePicker(false);
-                    }}
-                    whileTap={{ scale: 0.95 }}
-                  >
-                    <X />
-                  </IconButton>
-                )}
-              </EnhancedFilterContainer>
-
-              {conversations.length > 0 ? (
-                <EnhancedConversationGrid id="conversation-grid">
-                  {conversations.map((conv, index) => {
-                    if (!conv) return null;
-                    const messageCount = conv.chatMessages?.totalCount || 0;
-                    return (
-                      <EnhancedConversationCard
-                        key={conv.id}
-                        onClick={() => loadConversation(conv.id)}
-                        initial={{ opacity: 0, x: -10 }}
-                        animate={{ opacity: 1, x: 0 }}
-                        transition={{
-                          duration: 0.2,
-                          delay: index * 0.03,
-                        }}
-                      >
-                        <ChatItemIcon>
-                          <MessageCircle />
-                        </ChatItemIcon>
-                        <ChatItemContent>
-                          <ChatItemTitle>
-                            {conv.title || "Untitled Chat"}
-                          </ChatItemTitle>
-                          <ChatItemMeta>
-                            <span>
-                              {formatDistanceToNow(new Date(conv.createdAt))}{" "}
-                              ago
-                            </span>
-                            {conv.creator?.email && (
-                              <>
-                                <span>·</span>
-                                <span>{conv.creator.email}</span>
-                              </>
-                            )}
-                          </ChatItemMeta>
-                        </ChatItemContent>
-                        {messageCount > 0 && (
-                          <MessageCountBadge>{messageCount}</MessageCountBadge>
-                        )}
-                      </EnhancedConversationCard>
-                    );
-                  })}
-                  <FetchMoreOnVisible
-                    fetchNextPage={handleFetchMoreConversations}
-                  />
-                </EnhancedConversationGrid>
-              ) : (
-                <EmptyStateContainer>
-                  <EmptyStateIcon>
-                    <MessageCircle />
-                  </EmptyStateIcon>
-                  <EmptyStateTitle>No chats yet</EmptyStateTitle>
-                  <EmptyStateDescription>
-                    Start a conversation with the AI to ask questions about this
-                    corpus.
-                  </EmptyStateDescription>
-                  <EmptyStateButton
-                    onClick={startNewChat}
-                    whileHover={{ scale: 1.02 }}
-                    whileTap={{ scale: 0.98 }}
-                  >
-                    <Plus size={18} />
-                    Start New Chat
-                  </EmptyStateButton>
-                </EmptyStateContainer>
-              )}
-            </motion.div>
+            <ConversationListView
+              conversations={conversations}
+              onLoadConversation={loadConversation}
+              onStartNewChat={startNewChat}
+              onFetchMore={handleFetchMoreConversations}
+              titleFilter={titleFilter}
+              onTitleFilterChange={setTitleFilter}
+              createdAtGte={createdAtGte}
+              onCreatedAtGteChange={setCreatedAtGte}
+              createdAtLte={createdAtLte}
+              onCreatedAtLteChange={setCreatedAtLte}
+            />
           )}
         </AnimatePresence>
       </ConversationIndicator>
 
       {/* Approval Overlay */}
-      <AnimatePresence>
-        {(() => {
-          if (!pendingApproval || !showApprovalModal) return null;
-          return (
-            <motion.div
-              style={{
-                position: "absolute",
-                top: 0,
-                left: 0,
-                right: 0,
-                bottom: 0,
-                backgroundColor: "rgba(0, 0, 0, 0.5)",
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "center",
-                zIndex: 1000,
-                padding: "1rem",
-              }}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-            >
-              <motion.div
-                style={{
-                  backgroundColor: "white",
-                  borderRadius: "12px",
-                  padding: "2rem",
-                  maxWidth: "500px",
-                  width: "100%",
-                  boxShadow:
-                    "0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)",
-                }}
-                initial={{ scale: 0.9, y: 20 }}
-                animate={{ scale: 1, y: 0 }}
-                exit={{ scale: 0.9, y: 20 }}
-              >
-                <div
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    gap: "0.75rem",
-                    marginBottom: "1.5rem",
-                  }}
-                >
-                  <AlertCircle size={24} style={{ color: "#f59e0b" }} />
-                  <h3
-                    style={{ margin: 0, fontSize: "1.25rem", fontWeight: 600 }}
-                  >
-                    Tool Approval Required
-                  </h3>
-                  <button
-                    style={{
-                      marginLeft: "auto",
-                      background: "transparent",
-                      border: "none",
-                      cursor: "pointer",
-                    }}
-                    onClick={() => setShowApprovalModal(false)}
-                  >
-                    <X size={20} />
-                  </button>
-                </div>
-
-                <div style={{ marginBottom: "1.5rem" }}>
-                  <p style={{ margin: "0 0 1rem 0", color: "#374151" }}>
-                    The assistant wants to execute the following tool:
-                  </p>
-                  <div
-                    style={{
-                      backgroundColor: "#f3f4f6",
-                      padding: "1rem",
-                      borderRadius: "8px",
-                      fontFamily: "monospace",
-                      fontSize: "0.875rem",
-                    }}
-                  >
-                    <div style={{ fontWeight: 600, marginBottom: "0.5rem" }}>
-                      Tool: {pendingApproval.toolCall.name}
-                    </div>
-                    {Object.keys(pendingApproval.toolCall.arguments).length >
-                      0 && (
-                      <div>
-                        <div
-                          style={{ fontWeight: 600, marginBottom: "0.25rem" }}
-                        >
-                          Arguments:
-                        </div>
-                        <pre style={{ margin: 0, whiteSpace: "pre-wrap" }}>
-                          {JSON.stringify(
-                            pendingApproval.toolCall.arguments,
-                            null,
-                            2
-                          )}
-                        </pre>
-                      </div>
-                    )}
-                  </div>
-                </div>
-
-                <div
-                  style={{
-                    display: "flex",
-                    gap: "1rem",
-                    justifyContent: "flex-end",
-                  }}
-                >
-                  <Button
-                    size="medium"
-                    onClick={() => sendApprovalDecision(false)}
-                    style={{
-                      backgroundColor: "#dc2626",
-                      color: "white",
-                      border: "none",
-                      display: "flex",
-                      alignItems: "center",
-                      gap: "0.5rem",
-                    }}
-                  >
-                    <X size={16} />
-                    Reject
-                  </Button>
-                  <Button
-                    size="medium"
-                    onClick={() => sendApprovalDecision(true)}
-                    style={{
-                      backgroundColor: "#059669",
-                      color: "white",
-                      border: "none",
-                      display: "flex",
-                      alignItems: "center",
-                      gap: "0.5rem",
-                    }}
-                  >
-                    <CheckCircle size={16} />
-                    Approve
-                  </Button>
-                </div>
-              </motion.div>
-            </motion.div>
-          );
-        })()}
-      </AnimatePresence>
+      <ApprovalModal
+        pendingApproval={pendingApproval}
+        show={showApprovalModal}
+        onHide={() => setShowApprovalModal(false)}
+        onDecision={sendApprovalDecision}
+      />
     </ChatContainer>
   );
 };
