+++ frontend/src/components/knowledge_base/document/styled/index.ts
@@ -1,0 +1,66 @@
+export { EmptyState, EmptyStateContainer } from "./EmptyStates";
+export {
+  HeaderContainer,
+  MetadataRow,
+  ContentArea,
+  MainContentArea,
+  SummaryContent,
+} from "./HeaderAndLayout";
+export {
+  KnowledgeLayerContainer,
+  VersionHistorySidebar,
+  VersionHistoryHeader,
+  MobileBackButton,
+  VersionList,
+  VersionItem,
+  KnowledgeContent,
+  KnowledgeHeader,
+  KnowledgeBody,
+  EditModeToolbar,
+  EditToolbarCloseButton,
+  MarkdownEditor,
+  CollapseSidebarButton,
+} from "./KnowledgeLayer";
+export { TabsColumn, TabButton, TabTooltip } from "./LeftSidebar";
+export {
+  PlaceholderBase,
+  PlaceholderItem,
+  DocumentLoadingContainer,
+  SummaryPlaceholder,
+  NotePlaceholder,
+  RelationshipPlaceholder,
+  LoadingPlaceholders,
+} from "./LoadingStates";
+export {
+  RelationshipPanel,
+  RelationshipCard,
+  RelationshipType,
+} from "./Relationships";
+export {
+  ResizeHandle,
+  ResizeHandleControl,
+  ResizeHandleButton,
+  WidthControlBar,
+  WidthControlMenu,
+  WidthControlToggle,
+  WidthMenuItem,
+  MenuDivider,
+  WidthButton,
+  AutoMinimizeToggle,
+} from "./ResizeControls";
+export {
+  ControlButtonGroupLeft,
+  ConnectionStatus,
+  ControlButtonWrapper,
+  ControlButton,
+  SlidingPanel,
+  ChatIndicator,
+  ControlButtonGroup,
+} from "./RightPanel";
+export {
+  TabBadge,
+  MobileTabBar,
+  MobileTab,
+  SidebarTabsContainer,
+  SidebarTab,
+} from "./SidebarTabs";
